added Payment.ts interface, refactored payment test page to MakePayment, made ViewPayments for Get testing later
