+++ trivia-front-end/src/components/interfaces/Payment.ts
@@ -1,0 +1,7 @@
+export interface Payment{
+    paymentId: number,
+    studentId: number,
+    courseId: number,
+    date: string,
+    status: boolean
+}
