+++ trivia-front-end/src/components/payment/ViewPayments.tsx
@@ -1,0 +1,11 @@
+
+
+function ViewPayments() {
+  return (
+    <div>
+      
+    </div>
+  )
+}
+
+export default ViewPayments
